VOLD: Master key decryption compatiblity
When the data partition is encrypted with Jellybean then updated to
Kitkat the master key can no longer be decrypted.

The crypto footer data structure that contains the master key was
changed from Jellybean to Kitkat.  This patch moves the master key and
salt to the correct location in the cyrpto footer data structure.

Change-Id: I0df3149281d6ff7be4777cade5035fb161b843e6
Signed-off-by: Mark Allyn <mark.a.allyn@intel.com>
Author: Norvin Shuster <norvin.shuster@intel.com>
Committer: Mark Allyn <mark.a.allyn@intel.com>